requirements.txt
fastapi
uvicorn
chromadb
openai
pandas
python-docx
pdfplumber
PyJWT 
langchain 
requests
beautifulsoup4 
newspaper3k 
psycopg2-binary
python-dotenv
langchain
pypdf
python-docx
tqdm
python-multipart
langchain-community>=0.0.26
langchain>=0.1.13
openai>=1.0.0
pydantic>=2.0
pydantic-settings>=2.0.0
python-jose
sentence-transformers
tiktoken
tf-keras
nltk
pyspellchecker
transformers
streamlit
plotly
streamlit-chat